Refactor create rentals and modify app.rb
