+++ src/app.rb
@@ -6,6 +6,7 @@ require_relative './student'
 require_relative './teacher'
 require_relative './create_contents/create_book'
 require_relative './create_contents/create_person'
+require_relative './create_contents/create_rental'
 class App
   def initialize(menu)
     @menu = menu
@@ -40,6 +41,7 @@ class App
 
   def list_rentals_by_person_id
     puts 'No person entry found' if @people.empty?
+    @menu.option_list
     puts 'Enter person ID: '
     id = gets.chomp.to_i
     @people.each do |person|
@@ -50,30 +52,6 @@ class App
         puts "Date: #{rental.date}, Book: \"#{rental.book.title}\" by #{rental.book.author} "
       end
     end
-    sleep 0.75
-    @menu.option_list
-  end
-
-  def create_rental
-    puts 'Select a book from the following list by number'
-    @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title}, Author: #{book.author}" }
-    book_id = gets.chomp.to_i
-    puts 'Invalid input!' if !(book_id.is_a? Integer) && book_id > @books.length
-
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index do |person, index|
-      puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    person_id = gets.chomp.to_i
-    puts 'Invalide input!' if !(person_id.is_a? Integer) && person_id > @people.length
-
-    print 'Please enter the date in this format [yyyy-mm-dd]: '
-    date = gets.chomp.to_s
-
-    rental = Rental.new(date, @books[book_id], @people[person_id])
-    @rentals << rental
-
-    puts 'Rental created successfully'
     sleep 0.75
     @menu.option_list
   end
